requirements.txt
fastapi[all]
uvicorn
python-multipart
Pillow
fastapi-cors
opencv-python
numpy